Replace SDL_GetTicks64 with steady_clock::now() + std::chrono

+++ source/game/entities/entity_background.hpp
@@ -22,7 +22,7 @@ namespace SDLGame {
 
     private:
 
-        void Tick(const double&) override {
+        void Tick( const uint64_t& ) override {
 
         }
